+++ mockserver-core/src/main/java/org/mockserver/authentication/AuthenticationHandler.java
@@ -2,6 +2,6 @@ package org.mockserver.authentication;
 
 import org.mockserver.model.HttpRequest;
 
-public interface ControlPlaneAuthenticationHandler {
+public interface AuthenticationHandler {
     boolean controlPlaneRequestAuthenticated(HttpRequest request);
 }