+++ mockserver-netty/src/main/java/org/mockserver/netty/MockServer.java
@@ -6,13 +6,14 @@ import io.netty.buffer.PooledByteBufAllocator;
 import io.netty.channel.ChannelOption;
 import io.netty.channel.WriteBufferWaterMark;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
-import org.mockserver.authentication.jwt.ControlPlaneJWTAuthenticationHandler;
+import org.mockserver.authentication.ChainedAuthenticationHandler;
+import org.mockserver.authentication.jwt.JWTAuthenticationHandler;
+import org.mockserver.authentication.mtls.MTLSAuthenticationHandler;
 import org.mockserver.configuration.ConfigurationProperties;
 import org.mockserver.lifecycle.ExpectationsListener;
 import org.mockserver.lifecycle.LifeCycle;
 import org.mockserver.log.model.LogEntry;
 import org.mockserver.logging.MockServerLogger;
-import org.mockserver.authentication.mtls.ControlPlaneMTLSAuthenticationHandler;
 import org.mockserver.mock.action.http.HttpActionHandler;
 import org.mockserver.proxyconfiguration.ProxyConfiguration;
 import org.mockserver.socket.tls.NettySslContextFactory;
@@ -128,10 +129,17 @@ public class MockServer extends LifeCycle {
         }
 
         NettySslContextFactory nettySslContextFactory = new NettySslContextFactory(mockServerLogger);
-        if (ConfigurationProperties.controlPlaneTLSMutualAuthenticationRequired()) {
-            httpState.setControlPlaneAuthenticationHandler(new ControlPlaneMTLSAuthenticationHandler(mockServerLogger, nettySslContextFactory.trustCertificateChain(ConfigurationProperties.controlPlaneTLSMutualAuthenticationCAChain())));
+        if (ConfigurationProperties.controlPlaneTLSMutualAuthenticationRequired() && ConfigurationProperties.controlPlaneJWTAuthenticationRequired()) {
+            httpState.setControlPlaneAuthenticationHandler(
+                new ChainedAuthenticationHandler(
+                    new MTLSAuthenticationHandler(mockServerLogger, nettySslContextFactory.trustCertificateChain(ConfigurationProperties.controlPlaneTLSMutualAuthenticationCAChain())),
+                    new JWTAuthenticationHandler(mockServerLogger, ConfigurationProperties.controlPlaneJWTAuthenticationJWKSource())
+                )
+            );
+        } else if (ConfigurationProperties.controlPlaneTLSMutualAuthenticationRequired()) {
+            httpState.setControlPlaneAuthenticationHandler(new MTLSAuthenticationHandler(mockServerLogger, nettySslContextFactory.trustCertificateChain(ConfigurationProperties.controlPlaneTLSMutualAuthenticationCAChain())));
         } else if (ConfigurationProperties.controlPlaneJWTAuthenticationRequired()) {
-            httpState.setControlPlaneAuthenticationHandler(new ControlPlaneJWTAuthenticationHandler(mockServerLogger, ConfigurationProperties.controlPlaneJWTAuthenticationJWKSource()));
+            httpState.setControlPlaneAuthenticationHandler(new JWTAuthenticationHandler(mockServerLogger, ConfigurationProperties.controlPlaneJWTAuthenticationJWKSource()));
         }
         serverServerBootstrap = new ServerBootstrap()
             .group(bossGroup, workerGroup)